Fixed audio panel bug
Fixed the bug where the audio panel settings were not being set correctly
across scenes. This fix also brought to light a few other bugs with how
the current UIManager object was being handled. It is now initialized
within the initial scene, and is not destroyed across scenes, so all menu
panels now should be placed in this global object.