Created bounds zones to destroy bullets.
Bounds zones are triggers attached to a child of the camera. Any objects
with a 'Beam' component are destroyed upon entering. This will need to
change to be any objects with a 'Projectile' class once it's implemented,
as well as recycle objects into an object pool rather than destroy them.