Minor reorginization.
Created a Managers folder to hold all manager/singleton type scripts (The
Helpers folder was getting crowded, and most of the scripts weren't
actually managers). Also ensured none of the singleton scripts used the
Start function. As a habit, I will now avoid Start in favor of either
Awake or OnLevelWasLoaded. If I need something to happen in an
initialization step, it will go into the Awake function. If I need
something to happen at the beginning of a scene/level, then I will use the
OnLevelWasLoaded function.