Worked on save settings some more.
The audio portion of the save settings seems to be working well,
currently, the high score and the volume settings are saved. I decided to
not save the mute setting as there was a bug introduced, and I didn't feel
it was worth saving the status of mute over wasting time debugging it.

The high score hasn't been checked, but it's saving a value of zero into
the file currently. There may need to be some work put into getting high
score working properly in engine.