Added Medal system.
Added a medaling system built off of the Pickup script used on powerups. A
medaling system is where there are medals that appear in the game, and for
each consecutive medal the player picks up, they get a score bonus, but
they also have a medal multiplier that multiplies how much each
consecutive medal is worth. Dying or missing a medal will reset the medal
multiplier to 1.