Fixed audio pop bug.
Basically, when the player died, if they chose to quit the game, then the
CloseAllMenus function was called. Well, this function call the Pause
function. This caused some odd behavior for the split second between that
pause call, and the scene switching, with caused all shooting enemies
onscreen to shoot their weapons simultaneously.

Just adding in a quick Pause after closing the menu fixed this issue.